refactor: 당첨 통계 모달창 독립적으로 view 분리
